Updated and implemented Inventory design / code
Old inventory worked off of a dictionary. This design was simple, but missed the fact that an inventory can have multiple stacks of the same type of an item. Updated IStorer (and therefore Inventory) to work of the premise of the main data type being a two dimensional array. Items are now index by row and column. This makes intuitive sense as an inventory is generally grid based. While confining IStorers to being somewhat grid based, this solution is good enough for the scope of the project.