Updated Design - Simplified ISeller and IDisplayableSeller
Made IDisplayableSeller simply inherit from IDisplayableStorer and ISeller, then removed unnecessary duplicate functionality. Now ISeller still has +BuyItem() but should be able to work on InventoryUI.

!! IMPORTANT NOTE !!
This is not a great solution. There was a flaw in the earlier design that should have been caught before work was done. However, since work has already been done, instead of spending additional time refactoring, this quick and dirty solution should suffice 90% of the way. Since it is not a deeply integeral part of the design, the trade off seems reasonable.